# api/swagger-carsdemo.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cars demo",
        "description": "CMesh API User Profile, previously known as user resume.\n",
        "version": "20200201",
        "contact": {
            "email": "caruizag@gmail.com"
        }
    },
    "host": "localhost:8080",
    "basePath": "/20200201",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/cars/{carId}": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Get user details per user Id",
                "description": "Get user details per user id \n",
                "operationId": "GetCar",
                "parameters": [
                    {
                        "$ref": "#/parameters/RequestFields"
                    },
                    {
                        "in": "path",
                        "name": "carId",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation\n",
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "400": {
                        "$ref": "#/responses/400"
                    },
                    "401": {
                        "$ref": "#/responses/401"
                    },
                    "404": {
                        "$ref": "#/responses/404"
                    },
                    "500": {
                        "$ref": "#/responses/500"
                    }
                }
            }
        },
        "/cars": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Get user details per user Id",
                "description": "Get user details per user id \n",
                "operationId": "ListCars",
                "parameters": [
                    {
                        "$ref": "#/parameters/RequestFields"
                    },
                    {
                        "$ref": "#/parameters/PaginationLimitQueryParam"
                    },
                    {
                        "$ref": "#/parameters/PaginationTokenQueryParam"
                    },
                    {
                        "$ref": "#/parameters/SortOrderQueryParam"
                    },
                    {
                        "$ref": "#/parameters/SortByQueryParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation\n",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/CarSummary"
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/responses/400"
                    },
                    "401": {
                        "$ref": "#/responses/401"
                    },
                    "500": {
                        "$ref": "#/responses/500"
                    }
                }
            },
            "post": {
                "tags": [
                    "cars"
                ],
                "summary": "Add a new Car",
                "description": "",
                "operationId": "AddCar",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "New car to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation\n",
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "400": {
                        "$ref": "#/responses/400"
                    },
                    "401": {
                        "$ref": "#/responses/401"
                    },
                    "404": {
                        "$ref": "#/responses/404"
                    },
                    "409": {
                        "$ref": "#/responses/409"
                    },
                    "500": {
                        "$ref": "#/responses/500"
                    }
                }
            },
            "put": {
                "tags": [
                    "cars"
                ],
                "summary": "Update car",
                "description": "",
                "operationId": "UpdateCar",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Car to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation\n",
                        "schema": {
                            "$ref": "#/definitions/Car"
                        }
                    },
                    "400": {
                        "$ref": "#/responses/400"
                    },
                    "401": {
                        "$ref": "#/responses/401"
                    },
                    "404": {
                        "$ref": "#/responses/404"
                    },
                    "500": {
                        "$ref": "#/responses/500"
                    }
                }
            }
        }
    },
    "definitions": {
        "CarSummary": {
            "type": "object",
            "description": "Main car object .\n",
            "required": [
                "plate",
                "make",
                "model"
            ],
            "properties": {
                "plate": {
                    "type": "string",
                    "description": "Car plate\n"
                },
                "make": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "typeOfUse": {
                    "type": "string"
                }
            }
        },
        "Car": {
            "type": "object",
            "description": "Main car object .\n",
            "required": [
                "plate",
                "make",
                "model"
            ],
            "properties": {
                "plate": {
                    "type": "string",
                    "description": "Car plate\n"
                },
                "make": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "typeOfUse": {
                    "type": "string"
                },
                "nationalKey": {
                    "type": "string"
                },
                "serialNumber": {
                    "type": "string"
                },
                "engineSerialNumber": {
                    "type": "string"
                },
                "comparny": {
                    "type": "string"
                },
                "timeCreated": {
                    "type": "string",
                    "format": "date-time"
                },
                "timeUpdated": {
                    "type": "string",
                    "format": "date-time"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Image"
                    }
                }
            }
        },
        "Image": {
            "type": "object",
            "description": "Object that holds an image definition",
            "required": [
                "name",
                "content"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "content": {
                    "type": "file"
                }
            }
        },
        "Error": {
            "type": "object",
            "description": "Internal error object model.",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "description": "A short error code that defines the error, meant for programmatic parsing."
                },
                "message": {
                    "type": "string",
                    "description": "A human-readable error string."
                }
            }
        }
    },
    "parameters": {
        "RequestFields": {
            "name": "fields",
            "in": "query",
            "type": "array",
            "collectionFormat": "multi",
            "description": "Partial response refers to an optimization technique offered\nby the RESTful web APIs to return only the information \n(fields) required by the client. In this mechanism, the client\nsends the required field names as the query parameters for\nan API to the server, and the server trims down the default\nresponse content by removing the fields that are not required\nby the client. The parameter used to control what fields to\nreturn should be a query string parameter called \"fields\" of\ntype array, provide the values as enums, and usecollectionFormat\n",
            "items": {
                "type": "string"
            }
        },
        "PaginationLimitQueryParam": {
            "description": "The maximum number of items to return in a paginated \"List\" call.\nExample: `500`\n",
            "in": "query",
            "maximum": 1000,
            "minimum": 1,
            "name": "limit",
            "type": "integer"
        },
        "PaginationTokenQueryParam": {
            "description": "The value of page to display.\n",
            "in": "query",
            "maxLength": 512,
            "minLength": 1,
            "name": "page",
            "type": "string"
        },
        "SortOrderQueryParam": {
            "description": "The sort order to use, either ascending (`ASC`) or descending (`DESC`). The DISPLAYNAME sort order\nis case sensitive.\n",
            "enum": [
                "ASC",
                "DESC"
            ],
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "type": "string"
        },
        "SortByQueryParam": {
            "description": "The field to sort by. You can provide one sort order (`sortOrder`). Default order for\nTIMECREATED is descending. Default order for DISPLAYNAME is ascending. The DISPLAYNAME\nsort order is case sensitive.\n\n**Note:** In general, some \"List\" operations (for example, `ListInstances`) let you\noptionally filter by Availability Domain if the scope of the resource type is within a\nsingle Availability Domain. If you call one of these \"List\" operations without specifying\nan Availability Domain, the resources are grouped by Availability Domain, then sorted.\n",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "type": "string"
        }
    },
    "responses": {
        "400": {
            "description": "A bad request was made.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "401": {
            "description": "Unauthorized.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "404": {
            "description": "Not found.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "409": {
            "description": "Conflict with resource.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "429": {
            "description": "Too Many Requests",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "500": {
            "description": "Internal server error.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "default": {
            "description": "Unknown Error",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        }
    }
}